extra/requirements.txt
dynamixel_sdk==3.7.31
matplotlib==3.5.2
numpy==1.23.4
numpy==1.24.3
opencv_contrib_python==4.7.0.72
openpyxl==3.0.10
pandas==1.4.4
Pillow==9.4.0
Pillow==9.5.0
pyserial==3.5
pyusb==1.2.1
PyYAML==6.0
PyYAML==6.0
scikit_learn==1.0.2
scipy==1.9.1
screeninfo==0.8.1
torch==1.12.1
torchvision==0.13.1
